RF: move girder specifics around
